requirements.txt
aiohttp==3.9.3
aiohttp_session==2.12.0
anthropic==0.25.9
appdirs==1.4.4
autocorrect==2.6.1
blessed==1.20.0
CacheControl==0.12.11
colorhash==2.0.0
cryptography==41.0.1
dataclasses_json==0.5.6
fire==0.4.0
GitPython==3.1.43
Gymnasium==0.26.3
humanhash3==0.0.6
marshmallow==3.20.2
mashumaro==3.5
matplotlib==3.5.1
nest_asyncio==1.6.0
nltk==3.8.1
numpy==1.23.5
openai==1.30.1
orjson==3.10.3
peewee==3.14.8
Pillow==10.3.0
protobuf==5.26.1
pygame==2.1.2
python_dateutil==2.9.0
PyYAML==6.0.1
ray==2.9.3
Requests==2.31.0
sparklines==0.4.2
tiktoken==0.4.0
time_machine==2.8.2
torch==1.12.0
tqdm==4.64.0
transformers==4.37.2
viztracer==0.15.6
watchdog==3.0.0